fix: update review changes and config
