+++ .gitignore
@@ -42,3 +42,4 @@ jspm_packages
 #ide
 .vscode
 .idea
+server/idp-metadata
