+++ config/default.js
@@ -2,4 +2,5 @@
 module.exports = {
 	passportFile : '/opt/gluu-server/etc/gluu/conf/passport-config.json',
 	saltFile : '/opt/gluu-server/etc/gluu/conf/salt',
+	timerInterval: 600000,
 }
